Reorganized files to make more sense. Added show/hide path bar setting and view menu item.
