+++ CodeEdit/Features/Editor/TabBar/Tabs/Tab/EditorTabBackground.swift
