+++ CodeEdit/Features/Editor/TabBar/Views/EditorTabBarDivider.swift
@@ -1,5 +1,5 @@
 //
-//  TabBarDivider.swift
+//  EditorTabBarDivider.swift
 //  CodeEdit
 //
 //  Created by Lingxi Li on 4/22/22.
@@ -8,7 +8,7 @@
 import SwiftUI
 
 /// The vertical divider between tab bar items.
-struct TabDivider: View {
+struct EditorTabDivider: View {
     @Environment(\.colorScheme)
     var colorScheme
 
@@ -32,7 +32,7 @@ struct TabDivider: View {
 }
 
 /// The top border for tab bar (between tab bar and titlebar).
-struct TabBarTopDivider: View {
+struct EditorTabBarTopDivider: View {
     @Environment(\.colorScheme)
     var colorScheme
 
@@ -47,38 +47,16 @@ struct TabBarTopDivider: View {
                     .opacity(colorScheme == .dark ? 0.80 : 0.02)
                     .frame(height: tabBarStyle == .xcode ? 1.0 : 0.8)
                 // Shadow of top divider in native style.
-                TabBarNativeShadow()
+                EditorTabBarNativeShadow()
             }
         }
-    }
-}
-
-/// The bottom border for tab bar (between tab bar and breadcrumbs).
-struct TabBarBottomDivider: View {
-    @Environment(\.colorScheme)
-    var colorScheme
-
-    @AppSettings(\.general.tabBarStyle)
-    var tabBarStyle
-
-    var body: some View {
-        Rectangle()
-            .foregroundColor(
-                tabBarStyle == .xcode
-                ? Color(nsColor: .separatorColor)
-                    .opacity(colorScheme == .dark ? 0.80 : 1)
-                : Color(nsColor: .black)
-                    .opacity(colorScheme == .dark ? 0.65 : 0.13)
-
-            )
-            .frame(height: tabBarStyle == .xcode ? 1.0 : 0.8)
     }
 }
 
 /// The divider shadow for native tab bar style.
 ///
 /// This is generally used in the top divider of tab bar when tab bar style is set to `native`.
-struct TabBarNativeShadow: View {
+struct EditorTabBarNativeShadow: View {
     let shadowColor = Color(nsColor: .shadowColor)
 
     var body: some View {
