--- CodeEdit/Features/Editor/TabBar/Views/TabBarXcode.swift
@@ -1,15 +1,0 @@
-//
-//  TabBarXcode.swift
-//  CodeEdit
-//
-//  Created by Lingxi Li on 4/30/22.
-//
-
-import SwiftUI
-
-/// This is the Xcode style background material for tab bar and breadcrumbs.
-struct TabBarXcodeBackground: View {
-    var body: some View {
-        EffectView(.contentBackground)
-    }
-}
