+++ CodeEdit/Features/Settings/Pages/GeneralSettings/Models/GeneralSettings.swift
@@ -21,6 +21,9 @@ extension SettingsData {
         /// The show live issues behavior of the app
         var showLiveIssues: Bool = true
 
+        /// Show editor path bar
+        var showEditorPathBar: Bool = true
+
         /// The show file extensions behavior of the app
         var fileExtensionsVisibility: FileExtensionsVisibility = .showAll
 
@@ -81,6 +84,10 @@ extension SettingsData {
             self.showLiveIssues = try container.decodeIfPresent(
                 Bool.self,
                 forKey: .showLiveIssues
+            ) ?? true
+            self.showEditorPathBar = try container.decodeIfPresent(
+                Bool.self,
+                forKey: .showEditorPathBar
             ) ?? true
             self.fileExtensionsVisibility = try container.decodeIfPresent(
                 FileExtensionsVisibility.self,
